LF-3363 - Does not include deleted tasks, move util to test util, check if management_plan is deleted
