+++ packages/api/src/util/copyCropPlan.js
@@ -499,171 +499,3 @@ export const getManagementPlanGroupTemplateGraph = (
     }),
   };
 };
-
-/**
- * Used for jest tests only generates a management plan group graph based on the provided parameters.
- *
- * The function constructs a management plan group graph object for testing equality between copies in jest tests
- *
- * @param {Object} managementPlanGraph - The management plan graph used as the template object.
- *
- * @returns {Object} - Only the 'keep' properties of model template mapping in the form of a management plan group graph object.
- */
-export const getBareBonesManagementPlan = (managementPlanGraph) => {
-  return {
-    ..._omit(managementPlanGraph, getPropertiesToDelete(ManagementPlanModel)),
-    crop_management_plan: {
-      ..._omit(
-        managementPlanGraph.crop_management_plan,
-        getPropertiesToDelete(CropManagementPlanModel),
-      ),
-      planting_management_plans: managementPlanGraph.crop_management_plan.planting_management_plans.map(
-        (plan) => {
-          return {
-            ..._omit(plan, getPropertiesToDelete(PlantingManagementPlanModel)),
-            //Future looking to allow location changing LF-3367
-            location_id: plan.location_id,
-            managementTasks: plan.managementTasks
-              ? plan.managementTasks.map((managementTask) => {
-                  return {
-                    ..._omit(managementTask, getPropertiesToDelete(ManagementTasksModel)),
-                    task: {
-                      ..._omit(managementTask.task, getPropertiesToDelete(TaskModel)),
-                      //Future looking to allow location changing LF-3367
-                      coordinates: managementTask.task.coordinates,
-                      pest_control_task: managementTask.task.pest_control_task
-                        ? {
-                            ..._omit(
-                              managementTask.task.pest_control_task,
-                              getPropertiesToDelete(PestControlTaskModel),
-                            ),
-                          }
-                        : null,
-                      irrigation_task: managementTask.task.irrigation_task
-                        ? {
-                            ..._omit(
-                              managementTask.task.irrigation_task,
-                              getPropertiesToDelete(IrrigationTaskModel),
-                            ),
-                          }
-                        : null,
-                      scouting_task: managementTask.task.scouting_task
-                        ? {
-                            ..._omit(
-                              managementTask.task.scouting_task,
-                              getPropertiesToDelete(ScoutingTaskModel),
-                            ),
-                          }
-                        : null,
-                      soil_task: managementTask.task.soil_task
-                        ? {
-                            ..._omit(
-                              managementTask.task.soil_task,
-                              getPropertiesToDelete(SoilTaskModel),
-                            ),
-                          }
-                        : null,
-                      soil_amendment_task: managementTask.task.soil_amendment_task
-                        ? {
-                            ..._omit(
-                              managementTask.task.soil_amendment_task,
-                              getPropertiesToDelete(SoilAmendmentTaskModel),
-                            ),
-                          }
-                        : null,
-                      field_work_task: managementTask.task.field_work_task
-                        ? {
-                            ..._omit(
-                              managementTask.task.field_work_task,
-                              getPropertiesToDelete(FieldWorkTaskModel),
-                            ),
-                          }
-                        : null,
-                      harvest_task: managementTask.task.harvest_task
-                        ? {
-                            ..._omit(
-                              managementTask.task.harvest_task,
-                              getPropertiesToDelete(HarvestTaskModel),
-                            ),
-                          }
-                        : null,
-                      cleaning_task: managementTask.task.cleaning_task
-                        ? {
-                            ..._omit(
-                              managementTask.task.cleaning_task,
-                              getPropertiesToDelete(CleaningTaskModel),
-                            ),
-                          }
-                        : null,
-                      locationTasks: managementTask.task.locationTasks
-                        ? managementTask.task.locationTasks.map((locationTask) => {
-                            return {
-                              //Future looking to allow location changing LF-3367
-                              location_id: locationTask.location_id,
-                            };
-                          })
-                        : null,
-                    },
-                  };
-                })
-              : null,
-            plant_task: plan.plant_task
-              ? {
-                  ..._omit(plan.plant_task, getPropertiesToDelete(PlantTaskModel)),
-                  task: {
-                    ..._omit(plan.plant_task.task, getPropertiesToDelete(TaskModel)),
-                    coordinates: plan.plant_task.task.coordinates,
-                    locationTasks: plan.plant_task.task.locationTasks
-                      ? plan.plant_task.task.locationTasks.map((locationTask) => {
-                          return {
-                            //Future looking to allow location changing LF-3367
-                            location_id: locationTask.location_id,
-                          };
-                        })
-                      : null,
-                  },
-                }
-              : null,
-            transplant_task: plan.transplant_task
-              ? {
-                  ..._omit(plan.transplant_task, getPropertiesToDelete(TransplantTaskModel)),
-                  task: {
-                    ..._omit(plan.transplant_task.task, getPropertiesToDelete(TaskModel)),
-                    coordinates: plan.transplant_task.task.coordinates,
-                    locationTasks: plan.transplant_task.task.locationTasks
-                      ? plan.transplant_task.task.locationTasks.map((locationTask) => {
-                          return {
-                            //Future looking to allow location changing LF-3367
-                            location_id: locationTask.location_id,
-                          };
-                        })
-                      : null,
-                  },
-                }
-              : null,
-            bed_method: plan.bed_method
-              ? {
-                  ..._omit(plan.bed_method, getPropertiesToDelete(BedMethodModel)),
-                }
-              : null,
-            container_method: plan.container_method
-              ? {
-                  ..._omit(plan.container_method, getPropertiesToDelete(ContainerMethodModel)),
-                }
-              : null,
-            broadcast_method: plan.broadcast_method
-              ? {
-                  ..._omit(plan.broadcast_method, getPropertiesToDelete(BroadcastMethodModel)),
-                }
-              : null,
-            row_method: plan.row_method
-              ? {
-                  ..._omit(plan.row_method, getPropertiesToDelete(RowMethodModel)),
-                }
-              : null,
-          };
-        },
-      ),
-    },
-  };
-};
